requirements.txt
Flask==0.12.2
Flask-Babel==0.11.2
Flask-SQLAlchemy==2.3.2
Flask-Migrate==2.1.1
PyJWT==1.5.3
Flask-Mail==0.9.1
celery==4.1.0
